refactor: extract one batching engine for both transaction import paths
